pyproject.toml
[tool.poetry]
name = "djangosite"
version = "0.1.0"
description = ""
authors = ["night <night@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
django = "^4.1.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-debug-toolbar = "^3.7.0"
django-ckeditor = "^6.5.1"
python-dotenv = "^0.21.0"
djoser = "^2.1.0"
pillow = "^9.2.0"
django-widget-tweaks = "^1.4.12"
celery = "^5.2.7"
django-redis = "^5.2.0"
pytest = "^7.2.0"
django-redis-cache = "^3.0.1"
coverage = "^6.5.0"
gunicorn = "^20.1.0"
django-split-settings = "^1.2.0"
flower = "^1.2.0"
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.1"
mixer = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
